Production/French/FR-EN/requirements.txt
discord.py>=2.3.0
tzdata>=2023.3